Check if a semaphore can be acquired before waiting on it
We need to check, if a semaphore can be acquired in fuse_sem_wait before
waiting for its condition "count_cond" to be signaled. Otherwise, in case the
semaphore is ready to be acquired when fuse_sem_wait is called, we might never
be woken up to acquire the semaphore.