Makes non-error debug message less confusing.
In the POSIX API, errno is only valid when an error actually occurs,
and zero is not a valid error code.  Hence, strerror() maps 0
to "Unknown error".  However, the FUSE messaging API has an always
present "error" field which is zero in the no-error case.  The
debugging messages shouldn't report this as "Unknown error".