Implement new style mount process on OS X
When mounting a FUSE volume on Linux, fusermount opens the FUSE device
and passes the file descriptor back to libfuse.

However, on OS X libosxfuse used to be responsible for locating a free
FUSE device, opening it, and then passing the file descriptor to
mount_osxfuse. For this to work libosxfuse needs to be aware of
low level implementation details like the number of registered FUSE
devices.

In order to decouple libosxfuse from the FUSE kernel extension the
mount process has been revised to match FUSE on Linux. mount_osxfuse
will locate a free FUSE device, open it, and then pass the file
descriptor back to libosxfuse.