Turn off readdir_ino "on" by default
readdir_ino was turned "on" by default on Apple platforms, and it does not
appear to be required. getcwd works just fine without it turned on and
there is supporting evidence from a couple of places
(https://sourceforge.net/p/fuse/mailman/message/29170050/)
that it is basically ignored.

If we find places that users do need it, they can turn it on like on any other
file system.

This gets rid of a significant lock in readdir, and improves performance of
readdir significantly.