Restore readdir() fix for non-zero offset
Under macOS, telldir() may return zero the first time it is called.
This is the reason the readdir callback of the loopback reference file
system used to call the filler function with a zero offset for the
first directory entry and non-zero offsets for all subsequent directory
entries. This is a bug in the loopback file system.

Commit 06fc407 exposes this bug and readdir returns EIO. This lead to
falsely reverting 06fc407. This commit restores 06fc407.

This reverts commit 8071bd97494cd2982ed5635525b8a754edfbfb0f.